Split table into files
