+++ zkevm-circuits/src/table/rw_table.rs
@@ -1,0 +1,183 @@
+use super::*;
+
+/// Tag to identify the operation type in a RwTable row
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, EnumIter)]
+pub enum RwTableTag {
+    /// Start (used for padding)
+    Start = 1,
+    /// Stack operation
+    Stack,
+    /// Memory operation
+    Memory,
+    /// Account Storage operation
+    AccountStorage,
+    /// Tx Access List Account operation
+    TxAccessListAccount,
+    /// Tx Access List Account Storage operation
+    TxAccessListAccountStorage,
+    /// Tx Refund operation
+    TxRefund,
+    /// Account operation
+    Account,
+    /// Call Context operation
+    CallContext,
+    /// Tx Log operation
+    TxLog,
+    /// Tx Receipt operation
+    TxReceipt,
+}
+impl_expr!(RwTableTag);
+
+impl RwTableTag {
+    /// Returns true if the RwTable operation is reversible
+    pub fn is_reversible(self) -> bool {
+        matches!(
+            self,
+            RwTableTag::TxAccessListAccount
+                | RwTableTag::TxAccessListAccountStorage
+                | RwTableTag::TxRefund
+                | RwTableTag::Account
+                | RwTableTag::AccountStorage
+        )
+    }
+}
+
+impl From<RwTableTag> for usize {
+    fn from(t: RwTableTag) -> Self {
+        t as usize
+    }
+}
+
+/// The RwTable shared between EVM Circuit and State Circuit, which contains
+/// traces of the EVM state operations.
+#[derive(Clone, Copy, Debug)]
+pub struct RwTable {
+    /// Read Write Counter
+    pub rw_counter: Column<Advice>,
+    /// Is Write
+    pub is_write: Column<Advice>,
+    /// Tag
+    pub tag: Column<Advice>,
+    /// Key1 (Id)
+    pub id: Column<Advice>,
+    /// Key2 (Address)
+    pub address: Column<Advice>,
+    /// Key3 (FieldTag)
+    pub field_tag: Column<Advice>,
+    /// Key3 (StorageKey)
+    pub storage_key: Column<Advice>,
+    /// Value
+    pub value: Column<Advice>,
+    /// Value Previous
+    pub value_prev: Column<Advice>,
+    /// Aux1
+    pub aux1: Column<Advice>,
+    /// Aux2 (Committed Value)
+    pub aux2: Column<Advice>,
+}
+
+impl<F: Field> LookupTable<F> for RwTable {
+    fn columns(&self) -> Vec<Column<Any>> {
+        vec![
+            self.rw_counter.into(),
+            self.is_write.into(),
+            self.tag.into(),
+            self.id.into(),
+            self.address.into(),
+            self.field_tag.into(),
+            self.storage_key.into(),
+            self.value.into(),
+            self.value_prev.into(),
+            self.aux1.into(),
+            self.aux2.into(),
+        ]
+    }
+
+    fn annotations(&self) -> Vec<String> {
+        vec![
+            String::from("rw_counter"),
+            String::from("is_write"),
+            String::from("tag"),
+            String::from("id"),
+            String::from("address"),
+            String::from("field_tag"),
+            String::from("storage_key"),
+            String::from("value"),
+            String::from("value_prev"),
+            String::from("aux1"),
+            String::from("aux2"),
+        ]
+    }
+}
+impl RwTable {
+    /// Construct a new RwTable
+    pub fn construct<F: Field>(meta: &mut ConstraintSystem<F>) -> Self {
+        Self {
+            rw_counter: meta.advice_column(),
+            is_write: meta.advice_column(),
+            tag: meta.advice_column(),
+            id: meta.advice_column(),
+            address: meta.advice_column(),
+            field_tag: meta.advice_column(),
+            storage_key: meta.advice_column_in(SecondPhase),
+            value: meta.advice_column_in(SecondPhase),
+            value_prev: meta.advice_column_in(SecondPhase),
+            // It seems that aux1 for the moment is not using randomness
+            // TODO check in a future review
+            aux1: meta.advice_column_in(SecondPhase),
+            aux2: meta.advice_column_in(SecondPhase),
+        }
+    }
+    fn assign<F: Field>(
+        &self,
+        region: &mut Region<'_, F>,
+        offset: usize,
+        row: &RwRow<Value<F>>,
+    ) -> Result<(), Error> {
+        for (column, value) in [
+            (self.rw_counter, row.rw_counter),
+            (self.is_write, row.is_write),
+            (self.tag, row.tag),
+            (self.id, row.id),
+            (self.address, row.address),
+            (self.field_tag, row.field_tag),
+            (self.storage_key, row.storage_key),
+            (self.value, row.value),
+            (self.value_prev, row.value_prev),
+            (self.aux1, row.aux1),
+            (self.aux2, row.aux2),
+        ] {
+            region.assign_advice(|| "assign rw row on rw table", column, offset, || value)?;
+        }
+        Ok(())
+    }
+
+    /// Assign the `RwTable` from a `RwMap`, following the same
+    /// table layout that the State Circuit uses.
+    pub fn load<F: Field>(
+        &self,
+        layouter: &mut impl Layouter<F>,
+        rws: &[Rw],
+        n_rows: usize,
+        challenges: Value<F>,
+    ) -> Result<(), Error> {
+        layouter.assign_region(
+            || "rw table",
+            |mut region| self.load_with_region(&mut region, rws, n_rows, challenges),
+        )
+    }
+
+    pub(crate) fn load_with_region<F: Field>(
+        &self,
+        region: &mut Region<'_, F>,
+        rws: &[Rw],
+        n_rows: usize,
+        challenges: Value<F>,
+    ) -> Result<(), Error> {
+        let (rows, _) = RwMap::table_assignments_prepad(rws, n_rows);
+        for (offset, row) in rows.iter().enumerate() {
+            self.assign(region, offset, &row.table_assignment(challenges))?;
+        }
+        Ok(())
+    }
+}
